Move kotlin directory and fix alt
