+++ src/main/java/rocks/rdil/simpleconfig/Configuration.java
@@ -8,5 +8,5 @@ import java.lang.annotation.RetentionPolicy;
 @Target(ElementType.FIELD)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface Configuration {
-    String alt = "";
+    String alt();
 }
